# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "description": "Registers a user with default permissions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registers a user",
                "parameters": [
                    {
                        "description": "The user data to use when registering",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Token"
                            }
                        }
                    },
                    "400": {
                        "description": "Should only happen when there is already a user with the same username",
                        "schema": {}
                    },
                    "500": {
                        "description": "Shouldn't happen, but will if something fails",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/domain.GetUser"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Shouldn't happen, but will if something fails",
                        "schema": {}
                    }
                }
            }
        },
        "/api/v1/users/{userId}": {
            "get": {
                "description": "Get a user by id provided in the link",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The user id that you need to get",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetUser"
                        }
                    },
                    "400": {
                        "description": "Happens if the id isn't numerical",
                        "schema": {}
                    },
                    "404": {
                        "description": "Happens if there is no user with such id",
                        "schema": {}
                    },
                    "500": {
                        "description": "Shouldn't happen, but will if something fails",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreateUser": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "domain.GetUser": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "privilege": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "domain.Token": {
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        }
    }
}